requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
openai>=1.12.0
beautifulsoup4>=4.12.3
httpx>=0.27.0
redis>=5.0.1
python-dotenv>=1.0.1
tenacity>=8.2.3
google-cloud-pubsub>=2.19.0
google-cloud-firestore>=2.14.0
ddgs
sqlalchemy>=2.0.25
asyncpg>=0.29.0
greenlet>=3.0.3
trafilatura
python-dateutil
duckduckgo_search
curl_cffi
feedparser
aiosqlite
aiohttp
firebase-admin>=6.2.0
